# public/openapi.yaml
openapi: 3.0.0
info:
  title: Expense Tracker API
  description: A RESTful API for tracking personal expenses with user management and analytics
  version: 1.0.0
  contact:
    name: API Support
    email: support@expense-tracker.com
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.expense-tracker.com
    description: Production server
...
paths:
...
  /api/expenses:
    get:
      tags:
        - Expenses
      summary: Get all expenses
      description: Retrieve expenses with optional filters
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of expenses
          content:
            application/json:
              schema:
                type: object
                properties:
                  expenses:
                    type: array
                    items:
                      $ref: '#/components/schemas/Expense'
                  count:
                    type: integer

    post:
      tags:
        - Expenses
      summary: Create a new expense
      description: Add a new expense and trigger webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - userId
                - amount
                - category
              properties:
                userId:
                  type: integer
                amount:
                  type: number
                  format: float
                  example: 50.00
                category:
                  type: string
                  example: Food
                description:
                  type: string
                  example: Groceries
                date:
                  type: string
                  format: date
      responses:
        '201':
          description: Expense created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  expense:
                    $ref: '#/components/schemas/Expense'
        '400':
          description: Bad request
        '404':
          description: User not found
  /api/expenses/{userId}:
    get:
      tags:
        - Expenses
      summary: Get expenses for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: User expenses
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                  expenses:
                    type: array
                    items:
                      $ref: '#/components/schemas/Expense'
                  count:
                    type: integer
        '404':
          description: User not found
  /api/expenses/{userId}/total:
    get:
      tags:
        - Analytics
      summary: Get total expenses by category
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Total expenses by category
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseTotalResponse'
  /api/expenses/{userId}/monthly:
    get:
      tags:
        - Analytics
      summary: Get monthly spending summary
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Monthly spending breakdown
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlySpendingResponse'
...
components:
  schemas:
...
    Expense:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
          format: float
        category:
          type: string
        description:
          type: string
          nullable: true
        date:
          type: string
          format: date
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CategoryTotal:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
        total:
          type: number
        average:
          type: number
        minAmount:
          type: number
        maxAmount:
          type: number
        percentage:
          type: string
    ExpenseTotalResponse:
      type: object
      properties:
        userId:
          type: integer
        totalExpenses:
          type: number
        totalTransactions:
          type: integer
        byCategory:
          type: array
          items:
            $ref: '#/components/schemas/CategoryTotal'
    MonthlySpendingResponse:
      type: object
      properties:
        userId:
          type: integer
        totalExpenses:
          type: number
        totalTransactions:
          type: integer
        monthlyBreakdown:
          type: array
          items:
            type: object
        filters:
          type: object
          properties:
            year:
              type: string
            month:
              type: string
...